requirements.txt
fastapi==0.103.2
google==3.0.0
google-genai
PyMuPDF==1.25.4
python-dotenv
python-multipart
